+++ release_notes.md
@@ -1,3 +1,18 @@
+# 2.14.0
+
+- Allow defining transformation functions by points, not just record batches, for convenience.
+- Fix for bugs in domains.
+- Update dataset methods to allow instantiating from an arrow table represented
+  as a Uint8Array, which is simpler and avoids some conflicts in Apache Arrow versions.
+- Allow creating a dataset in the constructor, before the initial plotAPI call,
+  which allows more
+  precise staging of data and allows access to most dataset capabilities without binding to the DOM.
+- Support for Arrow Boolean types in transformations/selections.
+- New `Bitmask` type wrapping Boolean arrays on batches
+- Support combining of multiple selections into new selection using
+  fast `AND`, `OR`, `NOT`, and `XOR` operations.
+- Avoid race condition in multiple calls to the same selector.
+- Add applyToAllLoadedTiles method to selections.
 # 2.13.3
 
 Add `select_and_plot` shortcut on `Deepscatter` object to plot a selection immediately